refactor(startup): extract extension hook dispatch helpers into tau-onboarding
